requirements.txt
# fastapi>=0.111.0
# uvicorn[standard]>=0.30.0
# python-dotenv>=1.0.0
# python-multipart>=0.0.9
# pydantic>=2.7.0
# pydantic-settings>=2.3.0
# PyJWT>=2.8.0
# sqlalchemy>=2.0.30
# psycopg[binary]>=3.1.19
# pgvector>=0.3.0
# langchain>=0.2.0
# langchain-core>=0.2.0
# langchain-community>=0.2.0
# langgraph>=0.1.0
# langchain-google-genai>=1.0.0
# pypdf>=4.0.0
# passlib[bcrypt]>=1.7.4
# bcrypt>=4.0.0,<4.2.0
# fpdf2>=2.7.0
# google.generativeai==0.8.6
# grandalf
# duckduckgo-search>=6.0.0
# ddgs
# arize-phoenix>=4.0.0
# opentelemetry-sdk>=1.24.0
# openinference-instrumentation-langchain>=0.1.0


fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
python-multipart==0.0.20

pydantic==2.11.3
pydantic-settings==2.9.1

PyJWT==2.10.1

sqlalchemy==2.0.40
psycopg[binary]==3.2.6
pgvector==0.4.0

langchain==0.3.23
langchain-core==0.3.52
langchain-community==0.3.21
langgraph==0.3.34
langchain-google-genai==2.1.2

# google-generativeai==0.8.6

pypdf==5.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

fpdf2==2.8.3

duckduckgo-search==8.0.1
ddgs==9.5.2

arize-phoenix==8.24.0
opentelemetry-sdk==1.32.1
openinference-instrumentation-langchain==0.1.44

grandalf==0.8